requirements.txt
fastapi
uvicorn
jinja2
langchain
langchain-community
langchain-text-splitters
langgraph
python-dotenv
pandas
openpyxl
chromadb
